1nd-semester/LR5/python-grpc-glossary/client/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
flask==2.3.3